ti-j7xxx: trusted-firmware-a: update to fix CVE-2022-47630
Issue: LINUXEXEC-25581

This patch is to update trusted-firmware-a to fix the CVE-2022-47630, which will affect
trusted-firmware-a from v1.2 to v2.8,  and includes the following 4 patches:

- fd37982a19a4a291 "fix(auth): forbid junk after extensions"
- 72460f50e2437a85 "fix(auth): require at least one extension to be present"
- f5c51855d36e399e "fix(auth): properly validate X.509 extensions"
- abb8f936fd0ad085 "fix(auth): avoid out-of-bounds read in auth_nvctr()"

These patches have already been merged upstream.
And they will be deleted when trusted-firmware-a sync up to v2.8 or newer version.

(LOCAL REV: NOT UPSTREAM) - WRLinux Specific Patch

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>