ti-j7xxx: use KERNEL_DTC_FLAGS to pass dtc arguments instead of KERNEL_EXTRA_ARGS
Issue: LIN1023-2658

When DTC_FLAGS is passed to "make" via command line, it will overwrite the
kernel default value of DTC_FLAGS and trigger a lot of warnings. But when
DTC_FLAGS is passed via environment variable, Makefile of kernel can
handle it correctly by appending the default value to it.

For yocto, KERNEL_DTC_FLAGS is using environment to pass argument to
"make" and KERNEL_EXTRA_ARGS is using command line. So we should switch to
KERNEL_DTC_FLAGS.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>