ti-j7xxx: add ti-hsm-demo-fw to NON_MULTILIB_RECIPES
Issue: LIN1024-10256

The package "ti-hsm-demo-fw" is used for u-boot-ti-staging, it should not
build for lib32, so add it to NON_MULTILIB_RECIPES to fix below building
issue:

ERROR: lib32-ti-hsm-demo-fw-11.00.09-r4.0 do_deploy: Recipe lib32-ti-hsm-demo-fw is
trying to install files into a shared area when those files already exist. The files
and the manifests listing them are:
  tmp-glibc/deploy/images/ti-j78xx/ti-hsm/hsm-demo-firmware-j784s4-gp.bin
    (matched in manifest-ti_j78xx-ti-hsm-demo-fw.deploy)
  tmp-glibc/deploy/images/ti-j78xx/ti-hsm/hsm-demo-firmware-j784s4-hs-fs.bin
    (matched in manifest-ti_j78xx-ti-hsm-demo-fw.deploy)
  tmp-glibc/deploy/images/ti-j78xx/ti-hsm/hsm-demo-firmware-j784s4-hs.bin
    (matched in manifest-ti_j78xx-ti-hsm-demo-fw.deploy)
Please adjust the recipes so only one recipe provides a given file.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>